requirements.txt
asn1crypto==0.24.0
cffi==1.11.5
cryptography==2.4.2
Django==2.1.2
django-appconf==1.0.2
django-colorfield==0.1.15
django-filter==2.0.0
django-imagekit==4.0.2
django-rest-framework==0.1.0
djangorestframework==3.9.0
idna==2.7
pilkit==2.0
Pillow==5.3.0
psycopg2==2.7.5
pycparser==2.19
pytz==2018.7
six==1.11.0

